KH Classify: extract Sources tab into ClassifySourcesTabComponent
Source cards, source-detail slide-in (status filter, pagination, retry,
content items), and schedule dialog (save/remove/cron preview) moved to a
self-contained sub-page. Add/Edit/Run bubble up to host (form + pipeline stay
host-owned); DataChanged triggers source reload. Content-duplicates left in
host. CronToHumanReadable centralized in shared/classify.format (vector mgmt
import repointed). Host 4505 -> 3818 lines. Build green.
